requirements.txt
# Data Processing
kafka-python==2.0.2
paho-mqtt==1.6.1
pandas==2.0.3
numpy==1.24.3

# Database
TimescaleDB==2.11.0
psycopg2-binary==2.9.6

# Utilities
python-dotenv==1.0.0
loguru==0.7.0 